requirements.txt
cachetools==3.1.1
certifi==2020.6.20
chardet==3.0.4
futures==3.3.0
google-api-core==1.22.2
google-api-python-client==1.11.0
google-auth==1.21.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
oauthlib==3.1.0
Pillow==6.2.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-firebase==1.2
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.5
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.10
urwid==2.1.1